refactor: 커스텀 훅으로 수정, 아이템 데이터 관리
